+++ src/models/UserProfile.js
@@ -2,8 +2,19 @@
  * User profile factory class.
  */
 export class UserProfile {
-    name = "";
-    mail = "";
-    id = "";
+    name = null;
+    mail = null;
+    id = null;
     isAdmin = false;
+
+    constructor() {
+        this.name = "";
+        this.mail = "";
+        this.id = "";
+        this.isAdmin = false;
+    }
 }
+
+const userProfileInitialState = new UserProfile();
+
+export { userProfileInitialState };